requirements.txt
Pillow==7.0.0
pandas==1.0.5